requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
numpy>=1.24.0
faster-whisper>=0.10.0
requests>=2.31.0
pyyaml>=6.0
pydantic>=2.0.0
pybreaker>=1.0.0
prometheus-client>=0.19.0
